REFACTOR: Moved all the files related to grouping material into folder for grouping
